Removed base class, fixed oneturn config, removed 'small' try/excepts
