+++ pysixdesk/lib/preprocess.py
@@ -6,8 +6,8 @@ import argparse
 import time
 
 from pathlib import Path
+from contextlib import contextmanager
 
-from pysixdesk.lib.sixtrack_job import SixtrackJob
 from pysixdesk.lib import utils
 from pysixdesk.lib.dbtable import Table
 from pysixdesk.lib import generate_fort2
@@ -15,7 +15,7 @@ from pysixdesk.lib.pysixdb import SixDB
 from pysixdesk.lib.resultparser import parse_results
 
 
-class PreprocessJob(SixtrackJob):
+class PreprocessJob:
     def __init__(self, task_id, input_info):
         '''Class to handle the execution of the preprocessing job.
 
@@ -59,29 +59,171 @@ class PreprocessJob(SixtrackJob):
         # self.madx_cfg = cf['madx']
 
         output_files = self.madx_cfg["output_files"]
-        try:
-            output_files = utils.decode_strings(output_files)
-        except Exception:
-            content = "Wrong setting of madx output!"
-            raise ValueError(content)
+        output_files = utils.decode_strings(output_files)
         self.madx_out = output_files
         # self.mask_cfg = cf['mask']
 
-        self.six_cfg = cf['sixtrack']
-        self.fort_cfg = cf['fort3']
         self.oneturn_flag = self.madx_cfg.getboolean('oneturn')
         self.coll_flag = self.madx_cfg.getboolean('collimation')
 
-        # if self.oneturn_flag:
-        #     self.oneturn_cfg = cf['oneturn']
-        # else:
-        #     self.oneturn_cfg = {}
+        if self.oneturn_flag:
+            self.six_cfg = cf['sixtrack']
+            self.fort_cfg = cf['fort3']
+        else:
+            self.six_cfg = {}
+            self.fort_cfg = {}
         if self.coll_flag:
             self.coll_cfg = cf['collimation']
         else:
             self.coll_cfg = {}
         # close db to avoid unexpected timeouts
         self.db.close()
+
+    @contextmanager
+    def sixtrack_temp_folder(self, folder='temp'):
+        """Helper context manager to deal with the temp folder. On release,
+        moves back to the orignal folder and deletes the temp folder.
+
+        Args:
+            folder (str, optional): name of temp folder.
+        """
+        # quick context manager to handle the temporary folder
+        cwd = Path.cwd()
+        temp_folder = Path(folder)
+        if temp_folder.is_dir():
+            shutil.rmtree(temp_folder, ignore_errors=True)
+        temp_folder.mkdir()
+        try:
+            os.chdir(temp_folder)
+            yield
+        finally:
+            os.chdir(cwd)
+            shutil.rmtree(temp_folder, ignore_errors=True)
+
+    def sixtrack_prep_cfg(self, **kwargs):
+        """Prepares sixtrack's fort.3 config, by adding the length of the
+        machine, read in the fort.3.aux file, and any extra kwargs.
+
+        Args:
+            **kwargs: are added to the config.
+
+        Returns:
+            dict: fort.3 placeholder dictionnary with the added keys/values.
+        """
+        # make a dict copy of fort_cfg
+        fort_dic = dict(self.fort_cfg.items())
+        # reads the length from fort.3.aux
+        with open('fort.3.aux', 'r') as fc3aux:
+            fc3aux_lines = fc3aux.readlines()
+        fc3aux_2 = fc3aux_lines[1]
+        c = fc3aux_2.split()
+        lhc_length = c[4]
+        fort_dic['length'] = lhc_length
+        # adds additional kwargs to fort_dic
+        fort_dic.update(kwargs)
+        return fort_dic
+
+    def sixtrack_prep_job(self, fort_cfg, source_prefix=None,
+                          output_file='fort.3', symlink_parent=True):
+        """Prepares sixtrack fort.3 file and symlinks the other required
+        input files.
+
+        Args:
+            fort_cfg (dict): dict containing the placeholder/value pairs.
+            source_prefix (str/path, optional): if provided, will use the
+            provided folder prefix when looking for the fort_file and fc.3
+            files.
+            output_file (str, optional): name of the prepared fort.3 file.
+            symlink_parent (bool, optional): controls whether to symlink the
+            input_files of the parent folder to the current folder.
+
+        Raises:
+            Exception: If placeholder replacement in fort.3 fails.
+            FileNotFoundError: If missing input file during symlinking.
+            ValueError: If provided sixtrack 'input_files' are incorect.
+        """
+        # get the fort file patterns and values
+        if source_prefix is not None:
+            source_prefix = Path(source_prefix)
+        # touch fort.6
+        open('fort.6', 'a').close()
+
+        fort_file = Path(self.six_cfg["fort_file"])
+        patterns = ['%' + a for a in fort_cfg.keys()]
+        values = list(fort_cfg.values())
+        dest = fort_file.with_suffix('.temp')
+
+        source = fort_file
+        if source_prefix is not None:
+            source = source_prefix / source
+
+        utils.replace(patterns, values, source, dest)
+
+        # prepare the other input files
+        input_files = utils.decode_strings(self.six_cfg["input_files"])
+
+        madx_fc3 = input_files['fc.3']
+        if source_prefix is not None:
+            madx_fc3 = source_prefix / madx_fc3
+
+        # concatenate
+        utils.concatenate_files([dest, madx_fc3], output_file)
+        # if not source.samefile(output_file):
+        #     utils.diff(source, output_file, logger=self._logger)
+
+        # there could be merit in moving the symlinking to sixtrack_temp_folder
+        # so it would be done when entering a temp folder
+        if symlink_parent:
+            # make symlinks to the other input files in the parent folder
+            for file in [Path(f) for f in input_files.values()]:
+                target = Path.cwd().parent / file
+                if target.is_file():
+                    file.symlink_to(target)
+                else:
+                    raise FileNotFoundError("The required input file %s does not found!" %
+                                            file)
+
+    def sixtrack_copy_input(self):
+        '''
+        Copies the fort.3 file and other additional inputs to the cwd.
+
+        Args:
+            extra (list, optional): Any additional inputs to copy over.
+        '''
+        required = [self.six_cfg['fort_file']]
+        for infile in required:
+            infi = Path(self.six_cfg['source_path']) / infile
+            if infi.is_file():
+                shutil.copy2(infi, infile)
+            else:
+                content = "The required file %s isn't found!" % infile
+                raise FileNotFoundError(content)
+
+    def sixtrack_run(self, output_file):
+        """Runs sixtrack.
+
+        Args:
+            output_file (str): File in which to write sixtrack's stdout.
+        """
+        # actually run
+        six_output = os.popen(self.six_cfg["sixtrack_exe"])
+        self._logger.info('Sixtrack is running...')
+        # write stdout to file
+        outputlines = six_output.readlines()
+        self._logger.info('Sixtrack is done!')
+        # output_name = Path.cwd().parent / (job_name + '.output')
+
+        if outputlines and output_file is not None:
+            with open(output_file, 'w') as six_out:
+                six_out.writelines(outputlines)
+        elif output_file != 'fort.6':
+            # For some sixtrack version, the stdout will be automatically
+            # written to fort.6
+            shutil.copy2('fort.6', output_file)
+        # else:
+        #     self.output_files.append('fort.6')
+
+        # print(''.join(outputlines))
 
     def dl_output(self, dest_path):
         """Downloads the output of the job.
@@ -131,7 +273,18 @@ class PreprocessJob(SixtrackJob):
             val['task_id'] = [self.task_id] * len(val['mtime'])
             self.db.insertm(sec, val)
 
-        self.push_to_db_status(task_table, job_type='preprocess')
+        job_table = {}
+        if task_table['status'] == 'Success':
+            job_table['status'] = 'complete'
+            job_table['mtime'] = int(time.time() * 1E7)
+            self.db.update(f'preprocess_wu', job_table, f'task_id={self.task_id}')
+            content = f" preprocess task {self.task_id} has completed normally!"
+            self._logger.info(content)
+        else:
+            job_table['status'] = 'incomplete'
+            job_table['mtime'] = int(time.time() * 1E7)
+            self.db.update(f'preprocess_wu', job_table, f'task_id={self.task_id}')
+            self._logger.warning("This is a failed job!")
 
     def run(self):
         '''
@@ -201,15 +354,8 @@ class PreprocessJob(SixtrackJob):
         '''
         patterns = ['%' + a for a in self.mask_cfg.keys()]
         values = list(self.mask_cfg.values())
-        try:
-            utils.replace(patterns,
-                          values,
-                          self.madx_cfg["mask_file"],
-                          output_file)
-        except Exception:
-            content = "Failed to generate actual madx input file!"
-            raise Exception(content)
-
+        utils.replace(patterns, values, self.madx_cfg["mask_file"],
+                      output_file)
         # show diff
         # mask_name = self.madx_cfg["mask_file"]
         # utils.diff(mask_name, output_file, logger=self._logger)
@@ -308,7 +454,7 @@ class PreprocessJob(SixtrackJob):
                                    symlink_parent=True,
                                    output_file='fort.3')
             self.sixtrack_run(job_name)
-            # check and move output files
+            # check and move fort.10 file
             self.sixtrack_check(job_name)
 
     def sixtrack_job(self):
